# PERF: specialized 1D take version