@jreback 2021-03-05
> This gives a bit of code duplication (but all large parts of the function were already factored out into shared helper functions before, like _take_preprocess_indexer_and_fill_value), but having a 1D version gives less overhead compared to take_nd.

can you factor even more, this ultimately should be *used* in the other functions.