@jorisvandenbossche 2021-03-05
> this ultimately should be used in the other functions.

It *is* used in other functions, eg in unstack (added here, and now the other PR is merged, can also use it in reindex)