@jreback 2021-03-05
> pandas/core/internals/array_manager.py:997: error: Name 'take_nd' is not defined  [name-defined]