@jbrockmendel 2021-03-05
@jorisvandenbossche pls resist the temptation to self-merge.  i'll sleep better.

(no complaints about the PRs you've merged recently, just as you've commented elsewhere about premature merging, better to have good habits and stay on the safe side)

can this new function be used anywhere else, e.g. in Series/Index?