@jbrockmendel 2021-03-08
>  in case you think I made a wrong judgement call

All the case so far i think have been benign.  

> have been reviewed and are approved

In practice I doubt it will make a difference, but I will be moderately more comfortable if you ping me/jeff.  On the other hand if we're not reasonably responsive (timezones notwithstanding), do let me know.

Really the failure mode that I'm most worried about is something slipping through the cracks of my inbox.